Change prediction function to compute closest word_embedding
Because the output is not a one-hot vector anymore, the previous
prediction method does not work anymore. I changed it to compare the
output with all word-embeddings, the prediction is the embeddings that
has the shortest distance to the current output.

This does seem to work to some extent (init_word_embeddings.py), but
when the training set is increased, the network is not at all able to
predict anymore, even though the sum squared error still goes to 0.

I do not really understand why this is the case. The only thing I can
think of is that if the training set is limited, also the set of
embeddings that is actually used is limited, and the network can thus
keep them further apart, such that if the output is close to another, it
is far away from the others. I am not sure if this can be the cause, it
seems a bit unlikely, 10 dimensions should be more than enough to encode
around 140 different vectors.

Before further studying this issue, I will first implement the
make_batches function, so that the training data is properly batch
processed.