Fix prediction error batch SRN Theano
I fixed the batch computation of the prediction error in
generate_network_dynamics, at least on a syntactic level. The error
measure that is used now is the squared difference of the prediction (the
number) with the actual number. I did not test whether this works in
practice, which is the next step. For this I need to redefine the
computation of the prediction error in the 'test_single_sequence'
function, which I will do now.

I think I should also write a function that returns the errors for the
individual expressions, rather than only the sum squared error.