Add softmax classifier on top of network
Add softmax classifier to classify output of the network, so that cross
entropy can be used to train the network. The disadvantage of this
method is that the error mesarue does not reflect how 'close' the output
is, so to say, and it can thus not use this. Intuitively, the answer 6
should be better than the answer 15 if the true answer is 7. I should
maybe think about how to incorporate this.

NB: I tested whether this works with the file test_srn_theano.py.
Initially, I thought it didn't work, but this was caused by a too small
learning rate, and the network got stuck in a local minimum.