Create file to generate training data
generate_training_data uses class MathTreebank to generate
training data.
Currently the file contains a function to parse language names
such that examples with the right features can be generated, and
a function that generates a treebank object with the required number
of example sentences.
I should still write a function that converts this into numpy arrays,
based on information about the initial word embeddings.