Add functionality to monitor prediction error
I added a mean_squared_prediction_error function to the keras library,
that computes the difference between the (rounded) predicted outcome of
the input arithmetic expression and its true outcome. Add this function
to the metrics monitored when the model is compiled.
I added a function to plot the prediction error to the Training class.
I removed the original prediction function in the Training class.
I added a function to save the model to a file.