Add functionality to compute accuracy of input model on testset
I also made a start with computing the correlation between hidden layer
units, but intil now I did not manage to return the entire sequence
of activations of the recurrent layer (as opposed to only the last one
for every input sequence). I am not sure how to do this, maybe with a
truncated model?