Fix plots
The not plotting was due to using the wrong backend for matplotlib. This
is a little bit tricky because on the server the backend that is needed
to plot seems to cause an error with Tkinder and once the backend for
matplotlib is set it cannot be changed anymore it seems, which means
that in order for this file to work also the backend in the files
containing the imported classes should be changed (even though the
imports weren't directly imported). I should find a solution for this.